+++ src/state/UserProvider.js
@@ -1,10 +1,4 @@
-import React, {
-    useState,
-    createContext,
-    useContext,
-    useRef,
-    useEffect,
-} from 'react';
+import React, { useState, createContext, useContext } from 'react';
 
 const UserContext = createContext();
 
@@ -26,27 +20,3 @@ export const useMyArt = () => {
     const { myArt, setMyArt } = useContext(UserContext);
     return { myArt, setMyArt };
 };
-
-export function useCanvas(draw) {
-    const canvasRef = useRef(null);
-
-    useEffect(() => {
-        const canvas = canvasRef.current;
-        const context = canvas.getContext('2d');
-        let frameCount = 0;
-        let animationFrameId;
-
-        const render = () => {
-            frameCount++;
-            draw(context, frameCount);
-            animationFrameId = window.requestAnimationFrame(render);
-        };
-        render();
-
-        return () => {
-            window.cancelAnimationFrame(animationFrameId);
-        };
-    }, [draw]);
-
-    return canvasRef;
-}
